binder: make sure todo lists are handled in-order
It is assumed that the work items on the thread and proc todo
lists are handled in order. Since multiple threads could be
looking at the lists, the current thread handling the work
item at the head of the list freezes the list to prevent other
threads from dequeuing items.

Bug: 33250092 32225111
Change-Id: Iacd47a083a8d0af61a1287d79f28bd2419390909
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: DarkKnight6499 <yazad.madan@gmail.com>