binder: fix race condition between delete/add new ref on node
There was a race where 1 thread removed the final reference
on a node while a 2nd thread was adding a reference:

1. ThreadA get node to add reference
2. ThreadB delete last reference on node
3. node becomes zombie since no outstanding refs
4. ThreadA adds reference to zombie node (failure)

This is fixed by bumping the node's local_strong_refs
at step 1 so step 2 doesn't trigger deleting the node.
This implicit ref is taken on every call to binder_get_node()
or binder_new_node(), so the new binder_put_node() is added
which must be called when that node is no longer needed.

Bug: 34934663
Change-Id: Ib9c9dd74d2c2e9678f3fad084fb073b83abd1504
Test: EGL swapchain tests pass reliably
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: DarkKnight6499 <yazad.madan@gmail.com>